feat: convert into a library
